componentazation, cleanup, refactor, render optimization
